refactor[entity changes]: modified few entity data types, field names, constraints

+++ src/main/java/com/praveen/jpa/dao/OrderRepository.java
@@ -9,5 +9,5 @@ import java.util.List;
 @Repository
 public interface OrderRepository extends JpaRepository<Order, Long> {
 
-  List<Order> findByCustomerId(Integer customerId);
+  List<Order> findByCustomerId(Long customerId);
 }
